+++ src/tests/unittests/utilities/test_visualization.py
@@ -1,6 +1,5 @@
 import unittest
 from unittest.mock import Mock, patch
-
 import matplotlib.pyplot as plt
 
 import qtt.utilities.visualization
@@ -31,18 +30,18 @@ class TestVerticalHorizontalLine(unittest.TestCase):
 
     def test_plot_vertical_line(self):
         mock = Mock()
-        with patch('matplotlib.pyplot.axvline', return_value=mock) as mock_axvline:
-            vline = qtt.utilities.visualization.plot_vertical_line(3., color='r', alpha=.1, label='hi')
+        mock_ax = Mock()
+        mock_ax.axvline = Mock(return_value=mock)
+        vline = qtt.utilities.visualization.plot_vertical_line(3., color='r', alpha=.1, label='hi', ax=mock_ax)
 
-            mock_axvline.assert_called_once_with(3., label='hi')
-
+        mock_ax.axvline.assert_called_once_with(3., label='hi')
         self.assertEqual(vline, mock)
 
     def test_plot_horizontal_line(self):
         mock = Mock()
-        with patch('matplotlib.pyplot.axhline', return_value=mock) as mock_axhline:
-            vline = qtt.utilities.visualization.plot_horizontal_line(3., color='r', alpha=.1, label='hi')
+        mock_ax = Mock()
+        mock_ax.axhline = Mock(return_value=mock)
+        vline = qtt.utilities.visualization.plot_horizontal_line(4., color='r', alpha=.1, label='hi', ax=mock_ax)
 
-            mock_axhline.assert_called_once_with(3., label='hi')
-
+        mock_ax.axhline.assert_called_once_with(4., label='hi')
         self.assertEqual(vline, mock)